+++ locations/spiders/kfc_in.py
@@ -6,7 +6,7 @@ from scrapy.spiders import CrawlSpider, Rule
 
 from locations.items import Feature
 from locations.pipelines.address_clean_up import merge_address_lines
-from locations.spiders.kfc import KFC_SHARED_ATTRIBUTES
+from locations.spiders.kfc_us import KFC_SHARED_ATTRIBUTES
 
 
 class KfcINSpider(CrawlSpider):
