+++ locations/spiders/kfc_it.py
@@ -3,7 +3,7 @@ import json
 from scrapy import Spider
 
 from locations.dict_parser import DictParser
-from locations.spiders.kfc import KFC_SHARED_ATTRIBUTES
+from locations.spiders.kfc_us import KFC_SHARED_ATTRIBUTES
 
 
 class KFCITSpider(Spider):
