+++ locations/spiders/rostics_ru.py
@@ -4,7 +4,7 @@ from scrapy.http import JsonRequest
 from locations.categories import Categories, apply_category, apply_yes_no
 from locations.dict_parser import DictParser
 from locations.hours import DAYS_EN, OpeningHours
-from locations.spiders.kfc import KFC_SHARED_ATTRIBUTES
+from locations.spiders.kfc_us import KFC_SHARED_ATTRIBUTES
 
 
 class RosticsRUSpider(scrapy.Spider):
